# {Core} Build and use command index