… src/azure-cli-core/azure/cli/core/_session.py
… src/azure-cli-core/azure/cli/core/cloud.py
… src/azure-cli-core/azure/cli/core/commands/__init__.py
… src/azure-cli-core/azure/cli/core/commands/constants.py
… src/azure-cli-core/azure/cli/core/extension/operations.py
… src/azure-cli-core/azure/cli/core/tests/test_command_registration.py