## Description

A command index is built and saved in file `.azure/commandIndex.json`. It contains a mapping from **top-level commands / command group names** to their **command module namespaces**.

Both built-in modules and extensions are included in the index:

```jsonc
{
    "version": "2.7.0",
    "cloudProfile": "latest",
    "commandIndex": {
        "acr": [
            "azure.cli.command_modules.acr"
        ],
        ...
        "network": [
            "azure.cli.command_modules.natgateway",
            "azure.cli.command_modules.network",
            "azure.cli.command_modules.privatedns"
        ],
        ...
        "sql": [
            "azure.cli.command_modules.sql",
            "azure.cli.command_modules.sqlvm"
        ],
        ...
        "timeseriesinsights": [
            "azext_timeseriesinsights"
        ]
    }
}

```

CLI checks if a command can be found from the index (either from built-in modules, extensions, or both):

- ✔ Yes, then only corresponding modules from the index are loaded. Then check if the loaded modules do have the command, in case of an outdated cache:
    -  ✔ Yes, continue parsing with only the loaded modules
    -  ❌ No, go to (*)
- ❌ (*) No, then all modules and extensions will be loaded and a new index will be built from the loaded command table

This significantly decreases the run time of commands, since unnecessary modules are not loaded anymore.

Together with other PRs, Azure CLI's run time are improved from 2.3s to 0.6s (`az version`, without the loading time of `pkg_resources`):

- **Fix built-in timer** (https://github.com/Azure/azure-cli/pull/13844): Previous it only takes `invoke` duration into account, while forgetting about the `init` duration. The timing logic is fixed to include both `init` and `invoke` duration. 

- **Delay unnecessary imports** (https://github.com/Azure/azure-cli/pull/13843): Previously during CLI initialization, importing `azure.core` and `msrestazure.tools` are very time-consuming, but actually unnecessary. This PR delays the import until `azure.core` and `msrestazure.tools` are actually used. 

- **Remove unnecessary imports of msrest**: Follow [EAFP](https://docs.python.org/3/glossary.html#term-eafp) (Easier to Ask for Forgiveness than Permission) style to remove unnecessary imports of msrest. Also see https://stackoverflow.com/a/11360880/2199657

- **Add `util` module for `az rest` and `az version`**: (https://github.com/Azure/azure-cli/pull/13783): Peel off `az rest` and `az version` from `resource` module. A new module `util` is created. Thus they don't carry the burden of mgmt SDK initialization, thus making them faster. 
 
## Testing Guide

### Test with the built-in timer

First time run:

```
az version --debug

No module found from index for 'version'
Discovered command modules: ['acr', ... 'util', 'vm']
Loaded module 'acr' in 0.008 seconds.
...
Loaded module 'storage' in 0.059 seconds.
Loaded module 'util' in 0.004 seconds.
Loaded module 'vm' in 0.020 seconds.
Loaded all modules in 0.988 seconds. (note: there's always an overhead with the first module loaded)
Loading extensions: ['interactive', 'resource-graph', 'support', 'timeseriesinsights']
...
Extensions directory: 'D:\cli\azure-cli-extensions\src\timeseriesinsights'
Loaded extension 'timeseriesinsights' in 0.027 seconds.
Index updated: {'acr': ['azure.cli.command_modules.acr'], ... 'timeseriesinsights': ['azext_timeseriesinsights']}

Command ran in 2.007 seconds (init: 0.082, invoke: 1.926)
```

Second time run:

```
az version --debug

Modules found from index for 'version': ['azure.cli.command_modules.util']
Loading command modules: ['util']
Loaded module 'util' in 0.005 seconds.
Loaded all modules in 0.005 seconds. (note: there's always an overhead with the first module loaded)

Command ran in 0.283 seconds (init: 0.075, invoke: 0.208)
```

### Test with PowerShell

Before:
```
> Measure-Command {az version}

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 2
Milliseconds      : 343
Ticks             : 23430150
TotalDays         : 2.71182291666667E-05
TotalHours        : 0.0006508375
TotalMinutes      : 0.03905025
TotalSeconds      : 2.343015
TotalMilliseconds : 2343.015
```

After:
```
> Measure-Command {az version}

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 0
Milliseconds      : 676
Ticks             : 6760268
TotalDays         : 7.82438425925926E-06
TotalHours        : 0.000187785222222222
TotalMinutes      : 0.0112671133333333
TotalSeconds      : 0.6760268
TotalMilliseconds : 676.0268
```

### Test with Linux

Before:
```
$ time python -m azure.cli version

real    0m0.708s
user    0m0.629s
sys     0m0.080s
```

After:
```
$ time python -m azure.cli version

real    0m0.335s
user    0m0.314s
sys     0m0.020s
```

### Profile the import time

#### `az version -h`

```
python -X importtime -m azure.cli version -h 2>perf.log
tuna .\perf.log
```

Before:

![image](https://user-images.githubusercontent.com/4003950/80830830-44d4ef80-8c1c-11ea-8bf7-3b181b3fc145.png)

After:

![image](https://user-images.githubusercontent.com/4003950/80830559-beb8a900-8c1b-11ea-91b5-b99a2e8746b3.png)

#### `az version`

```
python -X importtime -m azure.cli version 2>perf.log
tuna .\perf.log
```

Before:

![image](https://user-images.githubusercontent.com/4003950/81138565-041b0480-8f95-11ea-922a-f5c814d98f14.png)

After:

![image](https://user-images.githubusercontent.com/4003950/81138477-cc13c180-8f94-11ea-9a74-19c3191bb4e5.png)

## Remark

- `pkg_resources` is still causing significant delay (0.272s) during the package initialization. That's where the extra time came from, compared to the built-in timer. This needs to be addressed separately in https://github.com/Azure/azure-cli/issues/13293

- See [Command line and environment](https://docs.python.org/3/using/cmdline.html#id5) for the usage of `python -X importtime`

- To quit `tuna`, on Window press <kbd>Ctrl</kbd>+<kbd>Break</kbd>, on Linux press <kbd>Ctrl</kbd>+<kbd>C</kbd>. See this [stackoverflow thread](https://stackoverflow.com/questions/1364173/stopping-python-using-ctrlc)