@fengzhou-msft 2020-05-06
When an extension is removed, will the command index be updated or invalidated?