@jiasli 2020-05-06
This also improves the running time of CI. Previously **CLI Automation Full Test** takes 1h 4m 2s. 

https://github.com/Azure/azure-cli/runs/628276350

![image](https://user-images.githubusercontent.com/4003950/81157102-39394e00-8fb9-11ea-9eff-22af3bc1db3a.png)

Now it only takes 27m 36s.
https://github.com/Azure/azure-cli/runs/648374336