@jiasli 2020-05-06
> When an extension is removed, will the command index be updated or invalidated?

Yes. In such case, the second check will fail and the index will be refreshed. https://github.com/Azure/azure-cli/blob/90c4ba188c87bec8ee634c63748d1dc4fac2be3b/src/azure-cli-core/azure/cli/core/__init__.py#L413