@haroldrandom 2020-05-18
Where represent the 3rd point, `Remove unnecessary imports of msrest` ?