@haroldrandom 2020-05-18
How do we maintain the commandIndex.json? (and considering the features/objective we discussed at brain storm)