@jiasli 2020-05-19
> How do we maintain the commandIndex.json? (and considering the features/objective we discussed at brain storm)

We don't need to maintain it. It is transparent and gets updated totally **automatically**.