@jiasli 2020-06-09
The command loading debug output is overhauled:

Before:
```
Installed command modules ['acr', 'acs', 'advisor', 'ams', ...]
Loaded module 'acr' in 0.007 seconds.
Loaded module 'acs' in 0.029 seconds.
Loaded module 'advisor' in 0.004 seconds.
Event: CommandLoader.OnLoadCommandTable []
Loaded module 'ams' in 0.006 seconds.
Loaded module 'apim' in 0.004 seconds.
Loaded module 'appconfig' in 0.004 seconds.
Loaded module 'appservice' in 0.008 seconds.
Loaded module 'aro' in 0.012 seconds.
```

After:

```
Loading command modules:
Module                Load Time   Groups Commands
acr                       0.006       25      103
acs                       0.012        7       50
advisor                   0.003        3        6
ams                       0.006       19       86
apim                      0.004        1        8
...
Total (61)                0.623      823     3321
Loading extensions:
Extension             Load Time   Groups Commands Directory
application-insights      0.006        6       12 C:\Users\user1\.azure\cliextensions\application-insights
interactive               0.002        0        0 C:\Users\user1\.azure\cliextensions\interactive
mesh                      0.006       12       26 C:\Users\user1\.azure\cliextensions\mesh
virtual-wan               0.008       11       43 C:\Users\user1\.azure\cliextensions\virtual-wan
webapp                    0.004        4        7 C:\Users\user1\.azure\cliextensions\webapp
azure-firewall            0.009       15       53 D:\cli\azure-cli-extensions\src\azure-firewall
timeseriesinsights        0.005        9       24 D:\cli\azure-cli-extensions\src\timeseriesinsights
Total (7)                 0.039       57      165
Loaded 876 groups, 3486 commands.
Updated command index in 0.002 seconds.
Event: CommandInvoker.OnPreCommandTableT
```

When using the command index:
```
Loading command modules:
Module                Load Time   Groups Commands
util                      0.003        1        2
Total (1)                 0.003        1        2
Loaded 1 groups, 2 commands.
```
It also shows a summary of the loaded commands:
```
Loaded 876 groups, 3486 commands.
```