@jiasli 2020-06-11
For commands including network transactions, we need to tightly work with Azure Core team to improve the import time. Currently it takes about 500ms to load.

```
> Measure-Command {python -c "from azure.core.exceptions import AzureError"}

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 0
Milliseconds      : 592
Ticks             : 5922874
TotalDays         : 6.85517824074074E-06
TotalHours        : 0.000164524277777778
TotalMinutes      : 0.00987145666666667
TotalSeconds      : 0.5922874
TotalMilliseconds : 592.2874
```

Tracked by https://github.com/Azure/azure-sdk-for-python/issues/11274 and https://github.com/Azure/azure-sdk-for-python/issues/11276.